+++ central/detection/service/service_impl.go
@@ -98,7 +98,12 @@ func (s *serviceImpl) DetectBuildTime(ctx context.Context, req *apiV1.BuildDetec
 	}
 
 	img := types.ToImage(req.GetImage())
-	enrichResult, err := s.imageEnricher.EnrichImage(enricher.EnrichmentContext{NoExternalMetadata: req.GetNoExternalMetadata()}, img)
+
+	eCtx := enricher.EnrichmentContext{}
+	if req.GetNoExternalMetadata() {
+		eCtx.FetchOpt = enricher.NoExternalMetadata
+	}
+	enrichResult, err := s.imageEnricher.EnrichImage(eCtx, img)
 	if err != nil {
 		return nil, err
 	}
@@ -220,8 +225,10 @@ func (s *serviceImpl) DetectDeployTimeFromYAML(ctx context.Context, req *apiV1.D
 	}
 
 	eCtx := enricher.EnrichmentContext{
-		NoExternalMetadata: req.GetNoExternalMetadata(),
-		EnforcementOnly:    req.GetEnforcementOnly(),
+		EnforcementOnly: req.GetEnforcementOnly(),
+	}
+	if req.GetNoExternalMetadata() {
+		eCtx.FetchOpt = enricher.NoExternalMetadata
 	}
 
 	var runs []*apiV1.DeployDetectionResponse_Run
@@ -296,8 +303,10 @@ func (s *serviceImpl) DetectDeployTime(ctx context.Context, req *apiV1.DeployDet
 	}
 
 	enrichmentCtx := enricher.EnrichmentContext{
-		NoExternalMetadata: req.GetNoExternalMetadata(),
-		EnforcementOnly:    req.GetEnforcementOnly(),
+		EnforcementOnly: req.GetEnforcementOnly(),
+	}
+	if req.GetNoExternalMetadata() {
+		enrichmentCtx.FetchOpt = enricher.NoExternalMetadata
 	}
 
 	run, err := s.enrichAndDetect(ctx, enrichmentCtx, req.GetDeployment())